update font scale and added more relative font sizing

+++ Source/UI/M1DropdownButton.h
@@ -73,7 +73,7 @@ public:
         return *this;
     }
     
-    M1DropdownButton & withFontSize(double fontSize_) {
+    M1DropdownButton & withFontSize(float fontSize_) {
         fontSize = fontSize_;
         return *this;
     }
